requirements.txt
# Core dependencies
python-dotenv>=1.0.0
openai>=2.12.0
chromadb>=1.3.7

# Web application
streamlit>=1.52.2

# Data processing
# Note: pandas 3.0.0 requires numpy>=1.26.0
# Using pandas 2.2.2 with numpy 1.26.4 for better stability
pandas>=2.2.2,<3.0.0
numpy>=1.26.4,<2.0.0

# Kaggle API (for dataset download)
kaggle>=1.8.2

# Utilities
tqdm>=4.67.1
